Move storage test dependencies to `io.spine.dependency.storage`
Declare `H2`, `HsqlDb`, and `MySql` in a new `io.spine.dependency.storage`
package, matching the `storage/` directory they live in (previously they
declared `io.spine.dependency.test`). Update the importing build script and
the `PredefinedMapping` version-sync comments to the new coordinates.
