+++ buildSrc/src/main/kotlin/io/spine/dependency/storage/H2.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.dependency.test
+package io.spine.dependency.storage
 
 /**
  * The H2 Database Engine — a fast, in-memory/embedded SQL database used for exercising
