+++ buildSrc/src/main/kotlin/io/spine/dependency/storage/MySql.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.dependency.test
+package io.spine.dependency.storage
 
 /**
  * MySQL Connector/J — the official JDBC driver for MySQL.
