+++ jdbc-storage/build.gradle.kts
@@ -29,9 +29,9 @@ import io.spine.dependency.lib.Hikari
 import io.spine.dependency.lib.QueryDsl
 import io.spine.dependency.lib.Slf4J
 import io.spine.dependency.local.CoreJvm
-import io.spine.dependency.test.H2
-import io.spine.dependency.test.HsqlDb
-import io.spine.dependency.test.MySql
+import io.spine.dependency.storage.H2
+import io.spine.dependency.storage.HsqlDb
+import io.spine.dependency.storage.MySql
 import io.spine.dependency.test.Testcontainers
 import org.gradle.api.tasks.testing.Test
